+++ info.rkt
@@ -19,5 +19,5 @@
                                     "gui-doc"))
 (define scribblings '(("scribblings/sauron.scrbl" (multi-page) (tool))))
 (define pkg-desc "A Racket IDE")
-(define version "1.0.4")
+(define version "1.1.0")
 (define pkg-authors '(dannypsnl))